requirements.txt
azure-data-tables
azure-core
